musicgen-api/requirements.txt
torch
transformers
fastapi
uvicorn
scipy
diffusers
accelerate
safetensors
pillow
